+++ src/domains/mod.rs
@@ -438,6 +438,7 @@ pub(crate) mod commodities;
     feature = "alphavantage",
     feature = "binance",
     feature = "crypto",
+    feature = "defi",
     feature = "fmp",
     feature = "kraken",
     feature = "polygon"
@@ -476,6 +477,7 @@ pub use commodities::Commodity;
     feature = "alphavantage",
     feature = "binance",
     feature = "crypto",
+    feature = "defi",
     feature = "fmp",
     feature = "kraken",
     feature = "polygon"