+++ src/models/mod.rs
@@ -41,6 +41,7 @@ pub mod commodities;
     feature = "alphavantage",
     feature = "binance",
     feature = "crypto",
+    feature = "defi",
     feature = "fmp",
     feature = "kraken",
     feature = "polygon"